+++ src/main/java/com/example/Lion.java
@@ -8,11 +8,8 @@ public class Lion {
 
     boolean hasMane;
 
-    public Lion(Feline feline) {
+    public Lion(Feline feline, String sex) throws Exception {
         this.feline = feline;
-    }
-
-    public Lion(String sex) throws Exception {
         if ("Самец".equals(sex)) {
             hasMane = true;
         } else if ("Самка".equals(sex)) {
