+++ src/test/java/FelineTest.java
@@ -1,37 +1,11 @@
 import com.example.Feline;
 import org.junit.Assert;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 
 import java.util.Arrays;
 import java.util.List;
 
-
-@RunWith(Parameterized.class)
 public class FelineTest {
-
-    private final int kittens;
-
-    public FelineTest(int kittens) {
-        this.kittens = kittens;
-    }
-
-    @Parameterized.Parameters
-    public static Object[] getData() {
-        return new Object[][] {
-                { 0},
-                { 1},
-                { 5}
-        };
-    }
-    @Test
-    public void getKittensByCount() {
-        Feline feline = new Feline();
-        int expectedResult = kittens;
-        int actualResult = feline.getKittens(kittens);
-        Assert.assertTrue(expectedResult == actualResult);
-    }
 
     @Test
     public void getKittens() {
@@ -50,14 +24,10 @@ public class FelineTest {
     }
 
     @Test
-    public void eatMeat() {
+    public void eatMeat() throws Exception {
         Feline feline = new Feline();
         List<String> expectedResult = Arrays.asList("Животные", "Птицы", "Рыба");
-        try {
-            List<String> actualResult = feline.eatMeat();
-            Assert.assertEquals(expectedResult, actualResult);
-        } catch (Exception e) {
-            System.out.println("Возникла непредвиденная ошибка");
-        }
+        List<String> actualResult = feline.eatMeat();
+        Assert.assertEquals(expectedResult, actualResult);
     }
 }
